myportal/requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.2.1
greenlet==2.0.2
idna==3.4
inflect==7.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
more-itertools==10.3.0
PyMySQL==1.1.1
python-dotenv==1.0.0
requests==2.31.0
sqlacodegen==2.3.0.post1
SQLAlchemy==1.4.52
typeguard==4.3.0
typing_extensions==4.12.2
urllib3==2.0.2
Werkzeug==2.3.4
WTForms==3.1.2
